requirements.txt
dnspython==2.1.0
flask==1.1.2
pymongo==3.11.3
python-dotenv==0.15.0
